requirements.txt
Django==1.10.5
Fabric3==1.13.1.post1
psycopg2==2.6.2
tweepy==3.5.0